+++ modules/speech_enhance/src/Beamformer.h
@@ -1,9 +1,8 @@
 #ifndef __BEAMFORMER_H__
 #define __BEAMFORMER_H__
 
-#include "basic_def.h"
-#include "basic_op.h"
-#include "complex.h"
+#include <stdint.h>
+#include <complex.h>
 
 typedef struct _Beamformer {
     uint32_t fn;
@@ -18,15 +17,16 @@ typedef struct _Beamformer {
 	complex float* steering;
 } Beamformer;
 
+
 // Beamformer
 int32_t Beamformer_Init(Beamformer* handle, uint32_t fftlen, uint32_t nchannel);
-void Beamformer_UpdateSteeringVector(Beamformer *handle, uint8_t update_speech, uint8_t update_noise);
-uint8_t Beamformer_UpdateSpeechMatrix(Beamformer* handle, complex float* X_itr, uint8_t speech_status);
-uint8_t Beamformer_UpdateNoiseMatrix(Beamformer* handle, complex float* X_itr, uint8_t noise_status,
+void Beamformer_UpdateSteeringVector(Beamformer *handle, uint32_t update_speech, uint32_t update_noise);
+uint32_t Beamformer_UpdateSpeechMatrix(Beamformer* handle, complex float* X_itr, uint32_t speech_status);
+uint32_t Beamformer_UpdateNoiseMatrix(Beamformer* handle, complex float* X_itr, uint32_t noise_status,
                                      float* spp);
-int32_t Beamformer_UpdateMvdrFilter(Beamformer* handle, uint8_t update_speech,
-                                    uint8_t update_noise);
-int32_t Beamformer_UpdateMwfFilter(Beamformer* handle, uint8_t update_speech, uint8_t update_noise);
+int32_t Beamformer_UpdateMvdrFilter(Beamformer* handle, uint32_t update_speech,
+                                    uint32_t update_noise);
+int32_t Beamformer_UpdateMwfFilter(Beamformer* handle, uint32_t update_speech, uint32_t update_noise);
 int32_t Beamformer_DoFilter(Beamformer* handle, complex float* X_itr, complex float* beamformed);
 int32_t Beamformer_Release(Beamformer* handle);
